+++ app/Http/Controllers/UsuarioController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\HistorialAccion;
+use App\Models\Obra;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -234,6 +235,19 @@ class UsuarioController extends Controller
     {
         DB::beginTransaction();
         try {
+            $usos = Obra::where("gerente_regional_id", $user->id)->get();
+            if (count($usos) > 0) {
+                throw ValidationException::withMessages([
+                    'error' =>  "No es posible eliminar este registro porque esta siendo utilizado por otros registros",
+                ]);
+            }
+            $usos = Obra::where("encargado_obra_id", $user->id)->get();
+            if (count($usos) > 0) {
+                throw ValidationException::withMessages([
+                    'error' =>  "No es posible eliminar este registro porque esta siendo utilizado por otros registros",
+                ]);
+            }
+
             $antiguo = $user->foto;
             if ($antiguo != 'default.png') {
                 \File::delete(public_path() . '/imgs/users/' . $antiguo);
@@ -256,10 +270,9 @@ class UsuarioController extends Controller
             ], 200);
         } catch (\Exception $e) {
             DB::rollBack();
-            return response()->JSON([
-                'sw' => false,
-                'message' => $e->getMessage(),
-            ], 500);
+            throw ValidationException::withMessages([
+                'error' =>  $e->getMessage(),
+            ]);
         }
     }
 }
